Remove duplicate devServer
The information for devServer config object is repeated twice. 

I also would personally would like to see the default port 3000 on the example - helps greatly when searching how to change this setting. Alternative option is to mention something in the line of `Change from default port 3000` in a comment line.