Fix corner case of hash scroll not happening (#462)
In my codebase that I did not manage to reduce for a demonstration,
RAF would delay the scrollTo call so much that the browser would have already instanly scrolled us to the correct location.
As a result the scrollTo would then be a no-op.

Instead, we now attempt to scroll as soon as possible, synchronously,
and fall back on RAF if that did not work.

Fixup efbaa1618. Fixes #449. Related to #460, but does not make that irrelevant.