Fix mocks to match with getConfig output
getConfig no longer needs to return devBasePath, stagingBasePath, or stagingSiteRoot.

Instead we resolve siteRoot and basePath to the correct values based on the environment inside getConfig function. For example if we are in staging, config.siteRoot will be stagingSiteRoot or siteRoot.

This allows us to remove a lot of conditionals around siteRoot and basePath scattered through the codebase.