Serve assets from its own location (#602)
* Add option to add `config.paths.assets`

* Update assets path with latest master changes

* Allow assetsPath to be absolute

* Add function to make a path absolute

* Add missing babel plugin

* Static data should be in assets too

* Add assetsPath and paths.assets to docs

* getSiteRoot was removed from buildXML

* Fix mocks to match with getConfig output

getConfig no longer needs to return devBasePath, stagingBasePath, or stagingSiteRoot.

Instead we resolve siteRoot and basePath to the correct values based on the environment inside getConfig function. For example if we are in staging, config.siteRoot will be stagingSiteRoot or siteRoot.

This allows us to remove a lot of conditionals around siteRoot and basePath scattered through the codebase.