New site using React-Static added to the list (#738)
Hi! Not sure if this is the right way to do this but we're excited to tell you that we've just relaunched or site after rebuilding it w/ React-Static! It's now super-fast and developing w/ it was 💪 and we couldn't be happier!

This PR/edit adds our site to your "Sites Built with React-Static" list.