Make sure plugin options are passed on to plugins
Options can be provided to plugins by passing an array instead of
the plugin name, with as first element the plugin name, and an
options object as the second argument. However, this array was
destructured as follows:

      originalLocation = originalLocation[0]
      options = originalLocation[1] || {}

Because the first line overwrote `originalLocation` with the name
of the plugin, the second line would actually assign the second
character of that name to `options`, rather than the second element
of the array. Switching them around fixes that.