Remove client from entry
According to the https://webpack.js.org/guides/hot-module-replacement/ documentation, the client should not be an entry.

It seems to be an old way of configuring (reference: article from 2016 https://medium.com/@rajaraodv/webpacks-hmr-react-hot-loader-the-missing-manual-232336dc0d96).

When this line is removed, the HMR plugin correct _injects_ this entry into the configuration.