Use browsers configuration from project
This removes the hardcoded browserlists by injecting them from the root path.

This is a "sorta breaking" change, as it changes the defaults from:
- >1%,
- last 4 versions,
- Firefox ESR,
- not ie < 9

To:

- > 0.5%
- last 2 versions
- Firefox ESR
- not dead

Since this is also a bug, it will be released with the minor release. Adding `browserslist` to your package.json or a config file, instantly fixes this.

#closes #1126