Bump packages and add devPeers
It's common to add peerDependencies as devDependencies. This way, they _will_ be installed on development and thus you can test as if they're really there.

Some of the peerDependencies were missing in the first place.

Additionally, during development, we want to use the local react-static folder module, which is symlinked.