Update README.md
I had several SSR issues when building the app with the latest `styled-components` 5.x.y installed. React Static documentation doesn't mention the version support, but looking at the Issues it's not compatible with the 5.x.y, so would be great to mention this on the documentation for people to avoid spending time on debugging. 

About the issue itself; for some reason, the CSS class names generated with `styled-components` 5.x.y installed clash with each other. This happens especially when extending components like so: 

```JS
const MyElement = styled.div`
  color: pink;
`;

const OtherElement = styled(MyElement)`
  background-color: yellow;
`;
```